requirements.txt
grpcio
protobuf
hstreamdb-api>=0.4.0,<0.5
